pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "github-research-feed"
version = "0.1.0"
description = "Personalized GitHub research intelligence feed as an MCP server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "anthropic>=0.40.0",
    "openai>=1.0.0",          # for text-embedding-3-small
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.20.0",
    "numpy>=1.26.0",
]

[project.scripts]
github-research-feed = "github_research_feed.server:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["github_research_feed*"]

[tool.setuptools.package-dir]
"github_research_feed" = "src/github_research_feed"
